+++ __tests__/networkViz.test.ts
@@ -1,19 +1,25 @@
 /**
  * @jest-environment jsdom
  */
+/// <reference types="jest" />
 
 import * as n from "../src/networkViz";
 import * as jsdom from 'jsdom';
 
-it("Should return Hello World", () => {
-    expect(n.networkViz.helloWorld()).toBe("Hello World")
+it("Should import as a function", () => {
+    var networkViz = require('../src/networkViz');
+    expect(networkViz).toHaveProperty("default");
+    expect(networkViz.default).toEqual(expect.any(Function));
 });
+
+
+
 
 
 /**
  * This is frankly incredible.
  */
-describe("Visit google.com", () => {
+describe("Click a button", () => {
     test("Go to google home page",  () => {
         const mockFn = jest.fn();
         document.body.innerHTML = `<div>